Modified loop() function to toggle sensors once, then motors.
TO TEST THIS CODE: input a HIGH signal to the masterTogglePin (3). This should collect sensor data once, then proceed to spin the motor clockwise.

Check if the direction of the motor is correct with respect to the sample collection assembly before testing! I'm not sure if it requires the motor to spin clockwise or anti.